progress2, just in case

+++ MandelbroSet/ComplexPlane.h
@@ -31,7 +31,7 @@ public:
 	void setMouseLocation(Vector2i mousePixel);
 	void loadText(Text& text);// might have to change we will see
 	void updateRender();
-	int countIterations(Vector2f coord);
+	size_t countIterations(Vector2f coord);
 	void iterationsToRGB(size_t count, Uint8& r, Uint8& g, Uint8& b);
 	Vector2f mapPixelToCoords(Vector2i mousePixel);
